+++ packages/client/src/layout/AppLayout.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import Home from '../home/';
+import Navbar from '../ui/Navbar';
+import Main from '../ui/Main';
+import Side from '../ui/Side';
+
+function AppLayout() {
+    return (
+        // <div>
+        //     <Navbar />
+        //     <Side />
+        //     <Main />
+        // </div>
+        <Home />
+    );
+}
+
+export default AppLayout;
