+++ packages/client/src/ui/Main.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Main() {
+    return <div>Main</div>;
+}
+
+export default Main;
