+++ packages/client/src/ui/Navbar.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Navbar() {
+    return <div>Navbar</div>;
+}
+
+export default Navbar;
