+++ packages/client/src/ui/ProtectedRoute.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function ProtectedRoute() {
+    return <div>ProtectedRoute</div>;
+}
+
+export default ProtectedRoute;
