+++ packages/client/src/ui/Side.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Side() {
+    return <div>Sidebar</div>;
+}
+
+export default Side;
